apollo_ar v0.7
added some things.

the bus routes are different depending on your direction, and now the
bus stop names and lat/longs account for that.

app never sleeps so you don't have to keep tapping the screen
(hopefully).

changed how the menu works a bit for touch screens.

and of course, the recordings of the street names are in place.